refactor: 북마크노트 null 반환시 예외처리 추가 #32

+++ src/main/java/com/main/writeRoom/service/NoteService/NoteCommandServiceImpl.java
@@ -235,6 +235,9 @@ public class NoteCommandServiceImpl implements NoteCommandService{
     @Transactional
     public void deleteBookmarkNote(Note note, User user) {
         BookmarkNote bookmarkNote = bookmarkNoteRepository.findByNoteAndUser(note, user);
+        if (bookmarkNote == null) {
+            throw new BookmarkHandler(ErrorStatus.BOOKMARK_NOT_FOUND);
+        }
         bookmarkNoteRepository.delete(bookmarkNote);
     }
 }
